requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-decouple==3.8
requests==2.28.2
django-cors-headers
django-utils-six
django-active-link
stripe
